lazy diagnostics
